GetPayDetail第一版，機車計費，未加月租抵扣及月租優惠費率

+++ WebAPI/Models/Param/Output/PartOfParam/MotorRentBase.cs
@@ -18,6 +18,6 @@ namespace WebAPI.Models.Param.Output.PartOfParam
         /// <summary>
         /// 每分鐘多少費用
         /// </summary>
-        public int MinuteOfPrice { set; get; }
+        public float MinuteOfPrice { set; get; }
     }
 }
